Adjust to size_t prototype

+++ lib/script.h
@@ -14,7 +14,7 @@
 
 extern const char *script_names[SCRIPT_NUM];
 
-extern int script_env_append(char ***, int *, const char *, const char *);
+extern int script_env_append(char ***, size_t *, const char *, const char *);
 extern void script_env_free(char **);
 extern int script_exec(const char *, int, char *const*);
